// src/app/models/art-initiation-demographics.ts
import { ArtInitiationDemographicEntry } from "./art-initiation-demographic-entry";

export class ArtInitiationDemographics {
  ageGroup_0_14: ArtInitiationDemographicEntry[] = [];
  ageGroup_15_24: ArtInitiationDemographicEntry[] = [];
  ageGroup_25_49: ArtInitiationDemographicEntry[] = [];
  ageGroup_50_plus: ArtInitiationDemographicEntry[] = [];
  }